YTCommentsSentimentAnalysis-main/requirements.txt
flask==2.3.3
google-api-python-client==2.97.0
pandas==2.0.3
openpyxl==3.1.2
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0